Silence compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@19005 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/Settings/SetOverlayPage.cpp
@@ -251,10 +251,10 @@ BOOL CSetOverlayPage::OnApply()
 		}
 		m_restart = Restart_None;
 	}
-	if (DWORD(m_regShowIgnoredOverlay) != m_bShowIgnoredOverlay)
+	if (DWORD(m_regShowIgnoredOverlay) != DWORD(m_bShowIgnoredOverlay))
 		m_restart = Restart_System;
 	Store (m_bShowIgnoredOverlay, m_regShowIgnoredOverlay);
-	if (DWORD(m_regShowUnversionedOverlay) != m_bShowUnversionedOverlay)
+	if (DWORD(m_regShowUnversionedOverlay) != DWORD(m_bShowUnversionedOverlay))
 		m_restart = Restart_System;
 	Store (m_bShowUnversionedOverlay, m_regShowUnversionedOverlay);
 	SetModified(FALSE);
